% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_max}
\alias{get_max}
\title{Maximum element from a vector}
\usage{
get_max(x)
}
\arguments{
\item{x}{vector of interest}
}
\value{
The maximum element from the vector of interest.
}
\description{
This function finds the maximum element from a vector.
}
