% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mode_factor}
\alias{mode_factor}
\title{Mode of a categorical variable}
\usage{
mode_factor(x)
}
\arguments{
\item{x}{categorical variable of interest.}
}
\value{
The mode of a categorical variable.
}
\description{
This function computes the mode of a categorical variable.
}
