% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{modefunc}
\alias{modefunc}
\title{Mode of a categorical variable}
\usage{
modefunc(x)
}
\arguments{
\item{x}{categorical variable of interest.}
}
\value{
the mode of a categorical variable
}
\description{
This function computes the mode of a categorical variable
}
