% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_importance.R
\name{relative_importance}
\alias{relative_importance}
\title{Relative importance of predictors in the model}
\usage{
relative_importance(Model, Data, Nrep = 10, r_seed = 1234)
}
\arguments{
\item{Model}{model of interest.}

\item{Data}{data frame used to fit the model of interest.}

\item{Nrep}{the number of times the randomisation process is repeated (default: 10).}

\item{r_seed}{the seed value (default: 1234).}
}
\value{
relImportance a vector providing mean relative importance, the lower bound of relative importance
(considering a 95\% confidence interval), median relative importance, and the upper bound of relative importance
(considenting a 95\% confidence interval).
}
\description{
This function evaluates the relative importance of predictors in a model. It implements the approach
developed by Thuiller et al. (2012), which involves comparing the predictions of a fitted model
(the "full model") with the predictions of models after randomisation of a given predictor, i.e. after random
permutation of the values of a given predictor within the dataset used for the fitted model ("random models").
One minus the Pearson’s correlation coefficient between the predictions of the full model and the predictions
of a random model gives an indication of the relative importance of a given predictor in explaining the
response variable of the full model. Note that the Pearson’s correlation coefficient can be negative.
Thuiller et al. (2012) consider that these cases represent an even greater importance of the permutated predictor
in explaining the response variable than with a correlation of 0.
}
\references{
Thuiller W, Lafourcade B, Araujo M. 2012. The presentation manual for BIOMOD. Grenoble, France:
Laboratoire d’Écologie Alpine, Université Joseph Fourier.
}
